Refactors alert into ducks
