--- client/actions/alert_actions.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const ALERT_RECIEVED = 'ALERT_RECEIVED';
-
-const ALERT_CLOSED = 'ALERT_CLOSED';
-
-
-module.exports = {
-  ALERT_RECIEVED,
-  ALERT_CLOSED,
-};
