+++ client/reducers/reducers.js
@@ -6,7 +6,7 @@ const rootReducer = combineReducers({
   //gacha: require('./gacha_reducer'),
   gacha: require('../modules/gacha').default,
   user: require('./user_reducer'),
-  alert: require('./alert_reducer'),
+  alert: require('../modules/alert').default,
   gossip: require('./gossip_reducer'),
   clock: require('./clock_reducer').clockReducer,
   dice: require('./dice_reducer'),
